Tidy up util classes for certificates
