+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/model/KafkaExporter.java
@@ -297,11 +297,11 @@ public class KafkaExporter extends AbstractModel {
      * @return The generated Secret.
      */
     public Secret generateCertificatesSecret(ClusterCa clusterCa, Secret existingSecret, boolean isMaintenanceTimeWindowsSatisfied) {
-        CertAndKey existingCertAndKey = CertUtils.keyStoreCertAndKey(existingSecret, COMPONENT_TYPE, clusterCa.caCertGenerationAnnotation());
+        CertAndKey existingCertAndKey = CertSecretUtils.keyStoreCertAndKey(existingSecret, COMPONENT_TYPE, clusterCa.caCertGenerationAnnotation());
 
         CertAndKey updatedCert = clusterCa.maybeCopyOrGenerateClientCert(reconciliation, componentName, existingCertAndKey, isMaintenanceTimeWindowsSatisfied);
 
-        Map<String, String> secretData = CertUtils.buildSecretData(COMPONENT_TYPE, updatedCert);
+        Map<String, String> secretData = CertSecretUtils.buildSecretData(COMPONENT_TYPE, updatedCert);
         return ModelUtils.createSecret(
                 KafkaExporterResources.secretName(cluster),
                 namespace,
