+++ operator-common/src/main/java/io/strimzi/operator/common/auth/PemTrustSet.java
@@ -7,6 +7,7 @@ package io.strimzi.operator.common.auth;
 import io.fabric8.kubernetes.api.model.Secret;
 import io.strimzi.operator.common.Util;
 import io.strimzi.operator.common.model.Ca;
+import io.strimzi.operator.common.model.CertificateUtils;
 
 import java.io.IOException;
 import java.security.GeneralSecurityException;
@@ -49,7 +50,7 @@ public class PemTrustSet {
                 .stream()
                 .map(cert -> {
                     try {
-                        return Ca.x509CertificateToPem(cert);
+                        return CertificateUtils.x509CertificateToPem(cert);
                     } catch (CertificateEncodingException e) {
                         throw new RuntimeException("Failed to convert X509 certificate to PEM format: " + cert.getSubjectX500Principal().getName(), e);
                     }
@@ -90,7 +91,7 @@ public class PemTrustSet {
                 .stream()
                 .map(entry -> {
                     try {
-                        return Ca.x509Certificate(entry.getValue());
+                        return CertificateUtils.x509Certificate(entry.getValue());
                     } catch (CertificateException e) {
                         throw new RuntimeException("Bad/corrupt certificate found in data." + entry.getKey() + " of Secret "
                                 + secretName + " in namespace " + secretNamespace);
