+++ operator-common/src/test/java/io/strimzi/operator/common/model/CaTest.java
@@ -21,7 +21,6 @@ import java.io.FileInputStream;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.security.cert.CertificateException;
-import java.security.cert.CertificateFactory;
 import java.security.cert.X509Certificate;
 import java.time.Clock;
 import java.time.Duration;
@@ -30,7 +29,6 @@ import java.time.ZonedDateTime;
 import java.time.temporal.ChronoUnit;
 import java.util.Base64;
 import java.util.Date;
-import java.util.List;
 import java.util.Map;
 
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -141,118 +139,15 @@ class CaTest {
             D0z+vgrfionoRhyWUDh7POlWwdUOWiBDBOFrkgeKNphSC0glYFN+2IW7
             -----END CERTIFICATE-----""";
 
-        X509Certificate x509 = Ca.x509Certificate(cert.getBytes());
+        X509Certificate x509 = CertificateUtils.x509Certificate(cert.getBytes());
         assertThat(x509.getSubjectX500Principal().getName(), is("CN=cluster-ca,O=Default Company Ltd,L=Default City,C=XX"));
 
-        String pem = Ca.x509CertificateToPem(x509);
+        String pem = CertificateUtils.x509CertificateToPem(x509);
         assertThat(pem, is(cert));
 
-        X509Certificate nextX509 = Ca.x509Certificate(pem.getBytes());
+        X509Certificate nextX509 = CertificateUtils.x509Certificate(pem.getBytes());
         assertThat(nextX509.getSubjectX500Principal().getName(), is("CN=cluster-ca,O=Default Company Ltd,L=Default City,C=XX"));
         assertThat(nextX509.getSignature(), is(x509.getSignature()));
-    }
-
-    @Test
-    @DisplayName("When certIsTrusted is called it correctly identifies whether the end entity certificate is issued by the CA cert")
-    public void testCertIsTrusted() throws IOException, CertificateException {
-        OpenSslCertIssuer ssl = new OpenSslCertIssuer();
-        CertificateFactory certFactory = CertificateFactory.getInstance("X.509");
-
-        File rootKey = createTempFile("key-", ".key");
-        File rootCert = createTempFile("crt-", ".crt");
-        File alternateRootKey = createTempFile("key-", ".key");
-        File alternateRootCert = createTempFile("crt-", ".crt");
-        File key = createTempFile("key-", ".key");
-        File csr = createTempFile("csr-", ".csr");
-        File cert = createTempFile("crt-", ".crt");
-
-        Subject rootSubject = new Subject.Builder().withCommonName("RootCn").withOrganizationName("MyOrganization").build();
-
-        // Generate a root cert
-        Instant now = Instant.now();
-        ZonedDateTime notBefore = now.truncatedTo(ChronoUnit.SECONDS).atZone(Clock.systemUTC().getZone());
-        ZonedDateTime notAfter = now.plus(2, ChronoUnit.HOURS).truncatedTo(ChronoUnit.SECONDS).atZone(Clock.systemUTC().getZone());
-        ssl.generateRootCaCert(rootSubject, rootKey, rootCert, notBefore, notAfter, 1);
-
-        // Generate alternate root cert
-        ssl.generateRootCaCert(rootSubject, alternateRootKey, alternateRootCert, notBefore, notAfter, 1);
-
-        Subject subject = new Subject.Builder()
-                .withCommonName("MyCommonName")
-                .withOrganizationName("MyOrganization")
-                .addDnsName("example1.com")
-                .addDnsName("example2.com").build();
-
-        // Generate cert
-        ssl.generateCsr(key, csr, subject);
-        ssl.generateCert(csr, rootKey, rootCert, cert, subject, 1);
-
-        X509Certificate x509RootCert = (X509Certificate) certFactory.generateCertificate(new FileInputStream(rootCert));
-        X509Certificate x509AlternateRootCert = (X509Certificate) certFactory.generateCertificate(new FileInputStream(alternateRootCert));
-        X509Certificate x509Cert = (X509Certificate) certFactory.generateCertificate(new FileInputStream(cert));
-
-        assertTrue(Ca.certIsTrusted(Reconciliation.DUMMY_RECONCILIATION, List.of(x509Cert), x509RootCert));
-        assertFalse(Ca.certIsTrusted(Reconciliation.DUMMY_RECONCILIATION, List.of(x509Cert), x509AlternateRootCert));
-    }
-
-    @Test
-    @DisplayName("When certIsTrusted is called it correctly identifies whether the end entity certificate is valid against a chain with intermediate certificates")
-    public void testCertChainWithIntermediateIsTrusted() throws IOException, CertificateException {
-        OpenSslCertIssuer ssl = new OpenSslCertIssuer();
-        CertificateFactory certFactory = CertificateFactory.getInstance("X.509");
-
-        File rootKey = createTempFile("key-", ".key");
-        File rootCert = createTempFile("crt-", ".crt");
-        File intermediateKey1 = createTempFile("key-", ".key");
-        File intermediateCert1 = createTempFile("crt-", ".crt");
-        File intermediateKey2 = createTempFile("key-", ".key");
-        File intermediateCert2 = createTempFile("crt-", ".crt");
-        File leafKey = createTempFile("key-", ".key");
-        File csr = createTempFile("csr-", ".csr");
-        File leafCert = createTempFile("crt-", ".crt");
-
-        Subject rootSubject = new Subject.Builder().withCommonName("RootCn").withOrganizationName("MyOrganization").build();
-
-        // Generate a root cert
-        Instant now = Instant.now();
-        ZonedDateTime notBefore = now.truncatedTo(ChronoUnit.SECONDS).atZone(Clock.systemUTC().getZone());
-        ZonedDateTime notAfter = now.plus(2, ChronoUnit.HOURS).truncatedTo(ChronoUnit.SECONDS).atZone(Clock.systemUTC().getZone());
-        ssl.generateRootCaCert(rootSubject, rootKey, rootCert, notBefore, notAfter, 1);
-
-        // Generate an intermediate cert
-        Subject intermediateSubject1 = new Subject.Builder().withCommonName("IntermediateCn1").withOrganizationName("MyOrganization").build();
-        ssl.generateIntermediateCaCert(rootKey, rootCert, intermediateSubject1, intermediateKey1, intermediateCert1, notBefore, notAfter, 1);
-
-        // Generate an additional intermediate cert
-        Subject intermediateSubject2 = new Subject.Builder().withCommonName("IntermediateCn2").withOrganizationName("MyOrganization").build();
-        ssl.generateIntermediateCaCert(intermediateKey1, intermediateCert1, intermediateSubject2, intermediateKey2, intermediateCert2, notBefore, notAfter, 1);
-
-
-        Subject subject = new Subject.Builder()
-                .withCommonName("MyCommonName")
-                .withOrganizationName("MyOrganization")
-                .addDnsName("example1.com")
-                .addDnsName("example2.com").build();
-
-        // Generate leaf cert
-        ssl.generateCsr(leafKey, csr, subject);
-        ssl.generateCert(csr, intermediateKey2, intermediateCert2, leafCert, subject, 1);
-
-        X509Certificate x509RootCert = (X509Certificate) certFactory.generateCertificate(new FileInputStream(rootCert));
-        X509Certificate x509IntermediateCert1 = (X509Certificate) certFactory.generateCertificate(new FileInputStream(intermediateCert1));
-        X509Certificate x509IntermediateCert2 = (X509Certificate) certFactory.generateCertificate(new FileInputStream(intermediateCert2));
-        X509Certificate x509LeafCert = (X509Certificate) certFactory.generateCertificate(new FileInputStream(leafCert));
-
-        assertFalse(Ca.certIsTrusted(Reconciliation.DUMMY_RECONCILIATION, List.of(x509LeafCert), x509RootCert));
-        assertFalse(Ca.certIsTrusted(Reconciliation.DUMMY_RECONCILIATION, List.of(x509LeafCert), x509IntermediateCert1));
-        assertTrue(Ca.certIsTrusted(Reconciliation.DUMMY_RECONCILIATION, List.of(x509LeafCert), x509IntermediateCert2));
-
-        assertTrue(Ca.certIsTrusted(Reconciliation.DUMMY_RECONCILIATION, List.of(x509LeafCert, x509IntermediateCert2), x509IntermediateCert1));
-        assertTrue(Ca.certIsTrusted(Reconciliation.DUMMY_RECONCILIATION, List.of(x509LeafCert, x509IntermediateCert2, x509IntermediateCert1), x509RootCert));
-
-        assertFalse(Ca.certIsTrusted(Reconciliation.DUMMY_RECONCILIATION, List.of(x509LeafCert, x509IntermediateCert1, x509IntermediateCert2), x509RootCert));
-        assertFalse(Ca.certIsTrusted(Reconciliation.DUMMY_RECONCILIATION, List.of(x509LeafCert, x509IntermediateCert2), x509RootCert));
-        assertFalse(Ca.certIsTrusted(Reconciliation.DUMMY_RECONCILIATION, List.of(x509IntermediateCert2, x509IntermediateCert1, x509LeafCert), x509RootCert));
     }
 
     @Test
